requirements.txt
azure-cognitiveservices-speech
edge_tts
python-socketio
requests
websocket-client